refactor: break up tui module into input, output, and mod
Split the monolithic tui.rs into a tui/ directory with three files:
- tui/mod.rs: App struct, render_app, event handling, run loop, tests
- tui/input.rs: AppState enum for input box state machine
- tui/output.rs: ConversationRole, ConversationEntry, conversation
  line building, truncation logic

Pure organizational refactor with no behavior changes.

Closes #63
