+++ src/frontend/tui/input.rs
@@ -1,0 +1,9 @@
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub enum AppState {
+    Input,
+    Streaming,
+    ToolConfirmation {
+        name: String,
+        input: serde_json::Value,
+    },
+}
